Move LambdaFacade and CognitoClient to services directory.
Update imports in various files.
